refactor, the continuation
