+++ Source/PluginProcessor.cpp
@@ -109,8 +109,8 @@ void SimpleEQAudioProcessor::prepareToPlay (double sampleRate, int samplesPerBlo
     
     updateFilters();
     
-    leftChannelFifo.prepare(samplesPerBlock);
-    rightChannelFifo.prepare(samplesPerBlock);
+    leftChannelProcessedBlockFifo.prepare(samplesPerBlock);
+    rightChannelProcessedBlockFifo.prepare(samplesPerBlock);
     
     osc.initialise([](float x) { return std::sin(x); });
     
@@ -188,10 +188,9 @@ void SimpleEQAudioProcessor::processBlock (juce::AudioBuffer<float>& buffer, juc
     
     leftChain.process(leftContext);
     rightChain.process(rightContext);
-    
-    leftChannelFifo.update(buffer);
-    rightChannelFifo.update(buffer);
-    
+
+    leftChannelProcessedBlockFifo.update(buffer);
+    rightChannelProcessedBlockFifo.update(buffer);
 }
 
 //==============================================================================
