+++ Source/Sound processing/Queue buffers/SingleChannelBlockFifoBuffer.cpp
@@ -1,0 +1,80 @@
+/*
+  ==============================================================================
+
+    SingleChannelBlockFifoBuffer.cpp
+    Created: 27 Feb 2023 9:56:56pm
+    Author:  digica_kwolk
+
+  ==============================================================================
+*/
+
+#include "SingleChannelBlockFifoBuffer.h"
+template<typename BlockType>
+struct SingleChannelBlockFifoBuffer
+{
+    enum Channel
+    {
+        Right, //effectively 0
+        Left //effectively 1
+    };
+
+    SingleChannelBlockFifoBuffer(Channel ch) : channelToUse(ch)
+    {
+        prepared.set(false);
+    }
+
+    void update(const BlockType& buffer)
+    {
+        jassert(prepared.get());
+        jassert(buffer.getNumChannels() > channelToUse);
+        auto* channelPtr = buffer.getReadPointer(channelToUse);
+
+        for (int i = 0; i < buffer.getNumSamples(); ++i)
+        {
+            pushNextSampleIntoFifo(channelPtr[i]);
+        }
+    }
+
+    void prepare(int bufferSize)
+    {
+        prepared.set(false);
+        size.set(bufferSize);
+
+        bufferToFill.setSize(1,             //channel
+            bufferSize,    //num samples
+            false,         //keepExistingContent
+            true,          //clear extra space
+            true);         //avoid reallocating
+        audioBufferFifo.prepare(1, bufferSize);
+        fifoIndex = 0;
+        prepared.set(true);
+    }
+    //==============================================================================
+    int getNumCompleteBuffersAvailable() const { return audioBufferFifo.getNumAvailableForReading(); }
+    bool isPrepared() const { return prepared.get(); }
+    int getSize() const { return size.get(); }
+    //==============================================================================
+    bool getAudioBuffer(BlockType& buf) { return audioBufferFifo.pull(buf); }
+private:
+    Channel channelToUse;
+    int fifoIndex = 0;
+    Fifo<BlockType> audioBufferFifo;
+    BlockType bufferToFill;
+    juce::Atomic<bool> prepared = false;
+    juce::Atomic<int> size = 0;
+
+    void pushNextSampleIntoFifo(float sample)
+    {
+        if (fifoIndex == bufferToFill.getNumSamples())
+        {
+            auto ok = audioBufferFifo.push(bufferToFill);
+
+            juce::ignoreUnused(ok);
+
+            fifoIndex = 0;
+        }
+
+        bufferToFill.setSample(0, fifoIndex, sample);
+        ++fifoIndex;
+    }
+};
